requirements.txt
transformers
datasets
pillow
numpy
torch
easyocr
reportlab
requests